refactor: moved 'types' into '/_core/' directory
