+++ src/js/initializers/init-burger-menu.ts
@@ -1,4 +1,4 @@
-import type { NullishHTMLElem } from '../types'
+import type { NullishHTMLElem } from '../_core/types'
 import BurgerMenu from '../modules/burger-menu'
 
 export default function initBurgerMenu(): void {
